test(scene-detail): split scene-detail specs by route, metadata, and recommendations
